+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: shikim <shikim@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/25 07:07:31 by shikim            #+#    #+#             */
-/*   Updated: 2023/04/25 07:07:38 by shikim           ###   ########.fr       */
+/*   Updated: 2023/04/25 08:20:34 by shikim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	length;
+	int	length;
 
 	length = 0;
 	while (s[length] != 0)
